1) Refactoring of arithmetic, logical and comparison operators. 2) Made operators work properly with all types, including strings and objects. 3) Merged all comparison operator handlers into one class to avoid complete code duplication.

+++ src/VM/Handlers/Opcode8045Handler.cpp
@@ -37,7 +37,7 @@ void Opcode8045Handler::_run()
 {
     Logger::debug("SCRIPT") << "[8045] [*] op_not" << std::endl;
     auto a = _vm->dataStack()->popLogical();
-    _vm->dataStack()->push(!a);
+    _vm->dataStack()->push((int)(!a));
 }
 
 }
